LightColoursFrame.java
When testing the GUI, I noticed if my selected light was off I was unable to change the light's color unless I used the Hue app or some type of voice assistant to turn it on. I added a few lines of code to test if the selected light is off and if so changes the selected light's state to on.